fix: renaming Rpc to RPC
